requirements.txt
unstructured
pandas
langchain
langchain_ollama
langchain-core
langchain-community
langchain-text-splitters
langchain-chroma
boto3
google-cloud-storage
azure-storage-blob
gdown
requests
rank_bm25
tabulate
streamlit
watchdog
fastapi
python-dotenv
dvc
docx
pdfplumber
langchain
pymongo
mlflow
python-pptx
python-multipart
plotly